Remove QueryValue.unapply and correct the numeric widening table (#321)
Both are source-breaking, so they are held for 2.0 rather than folded into the
compatible set.

QueryValue.unapply
------------------
The decode half of the typeclass -- thirteen implementations -- had no caller
anywhere in the repository except one internal recursive use, and was wrong:

  private def unquote(v: String) = v.substring(1, v.length - 2)

`quote` adds one character at each end, so this dropped one too many and 'abc'
came back as "ab". Nothing noticed because nothing called it.

Decoding query results is a different problem with a different shape -- results
are JSON, not SQL literals -- so the reverse direction is removed rather than
fixed. QueryValue is now the one-way serializer it always actually was.

Breaking: removes an abstract method from a public trait, so any third-party
QueryValue implementation stops compiling (deleting its unapply is the fix).

AritRetType
-----------
Three entries disagreed with ClickHouse. Checked with toTypeName on 25.3 rather
than reasoned about:

  toInt64(10) %  toInt32(3)   -> Int64     was AritRetType[Long, Int, Int]
  toFloat64(1) + toFloat32(2) -> Float64   was AritRetType[Double, Float, Float]
  toFloat32(1) + toFloat64(2) -> Float64   was AritRetType[Float, Double, Float]

The first was also asymmetric with IntLongBinding, which already widened
Int x Long to Long. The other two narrowed in both directions.

Breaking: the phantom result type of an arithmetic expression changes, so code
that spells it out explicitly no longer compiles. HigherOrderFunctionsIT is the
example -- arrayReverseSort[Long, Int](Some(_ % 3), ...) had been written to
match the narrowing and becomes [Long, Long].

Note that this table cannot be right for division: AritRetType is keyed only on
the operand types, but ClickHouse's `/` always yields Float64 (Int64 / Int32 is
Float64) while intDiv yields Int64. Distinguishing them needs the operator in
the key, which is left for the typed-results work.

+++ dsl/src/it/scala/com/crobox/clickhouse/dsl/column/HigherOrderFunctionsIT.scala
@@ -103,8 +103,10 @@ class HigherOrderFunctionsIT extends DslITSpec {
   }
 
   it should "HigherOrderFunctions: arrayReverseSort" in {
-    r(arrayReverseSort[Long, Int](Some(_ % 3), arr1)) shouldBe "[2,1,3]"
-    r(arrayReverseSort[Long, Int](None, arr1)) shouldBe "[3,2,1]"
+    // [Long, Long], not [Long, Int]: ClickHouse's `modulo(Int64, Int32)` is Int64 (verified with toTypeName), and the
+    // AritRetType table now agrees -- it used to narrow Long % Int to Int, which is what this line was written against.
+    r(arrayReverseSort[Long, Long](Some(_ % 3), arr1)) shouldBe "[2,1,3]"
+    r(arrayReverseSort[Long, Long](None, arr1)) shouldBe "[3,2,1]"
 
     r(arrayReverseSort2[Int, Int]((x, y) => x * y, Seq(1, 3, 3, 0), Seq(3, 2, 3, 4))) shouldBe "[3,3,1,0]"
   }
